refactor: utils 함수 domain/UI 분리
UI 부분 import 문을 필요로 하는 함수를 따로 분리
이를 통해 step2에서 readline을 불러오지 않을 수 있음
